backend/video/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.1
gevent==24.2.1
mysqlclient==2.2.1
Pillow==10.2.0
ffmpeg-python==0.2.0
python-dotenv==1.0.1
gunicorn==21.2.0
waitress==3.0.0
uvicorn[standard]==0.27.0
django-filter==23.5
drf-yasg==1.21.7
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0 